Changed the structure by breaking `base.py` into a directory with one class per file. Inner-scope imports remained, though.
